added ThreadPool public API
