+++ .gitignore
@@ -76,5 +76,8 @@ third_party/abseil-cpp
 third_party/benchmark
 
 python/sentencepiece
+python/test/m_*.model
+python/test/m_*.vocab
+python/test/sp_*.pickle
 
 third_party/absl
