+++ python/src/sentencepiece/sentencepiece_wrap.cxx
@@ -3849,31 +3849,6 @@ inline void ConvertToUnicodeSpans(sentencepiece::ImmutableNBestSentencePieceText
   proto->ConvertToUnicodeSpans();
 }
 
-class ThreadPool {
- public:
-  explicit ThreadPool(size_t request_size) :
-    request_size_(request_size) {}
-
-  virtual ~ThreadPool() {
-    for (auto &task : tasks_) {
-      task.join();
-    }
-  }
-
-  void Schedule(std::function<void()> closure) {
-    static constexpr size_t kMinThreadSize = 2;
-    if (request_size_ < kMinThreadSize) {
-      closure();
-    } else {
-      tasks_.emplace_back(closure);
-    }
-  }
-
- private:
-  size_t request_size_ = 0;
-  std::vector<std::thread> tasks_;
-};
-
 template <typename T>
 inline void InitNumThreads(const std::vector<T> &ins, int *num_threads) {
   if (*num_threads < 0) {
@@ -3888,7 +3863,7 @@ inline void InitNumThreads(const std::vector<T> &ins, int *num_threads) {
   std::vector<OutType> outs(ins.size());                                \
   InitNumThreads(ins, &num_threads);                                    \
   {                                                                     \
-    ThreadPool pool(ins.size());                                        \
+    sentencepiece::ThreadPool pool(num_threads);                        \
     std::atomic<size_t> index = 0;                                      \
     for (int n = 0;  n < num_threads; ++n) {                            \
       pool.Schedule([&]() {                                             \
@@ -3913,7 +3888,7 @@ inline void InitNumThreads(const std::vector<T> &ins, int *num_threads) {
   InitNumThreads(ins, &num_threads);                                    \
   {                                                                     \
     std::atomic<size_t> index = 0;                                      \
-    ThreadPool pool(ins.size());                                        \
+    sentencepiece::ThreadPool pool(num_threads);                        \
     for (int n = 0;  n < num_threads; ++n) {                            \
       pool.Schedule([&]() {                                             \
           size_t i = 0;                                                 \
@@ -4524,7 +4499,7 @@ SWIGINTERN std::vector< float > sentencepiece_SentencePieceProcessor__CalculateE
     std::vector<float> outs(ins.size());
     InitNumThreads(ins, &num_threads);
     {
-      ThreadPool pool(ins.size());
+      sentencepiece::ThreadPool pool(num_threads);
       std::atomic<size_t> index = 0;
       for (int n = 0;  n < num_threads; ++n) {
         pool.Schedule([&]() {
