+++ src/util.h
@@ -361,61 +361,6 @@ void STLDeleteElements(std::vector<T *> *vec) {
 }
 }  // namespace port
 
-class ThreadPool {
- public:
-  explicit ThreadPool(int num_threads) {
-    threads_.reserve(num_threads);
-    for (int i = 0; i < num_threads; ++i) {
-      threads_.push_back(std::thread(&ThreadPool::WorkLoop, this));
-    }
-  }
-
-  ThreadPool(const ThreadPool &) = delete;
-  ThreadPool &operator=(const ThreadPool &) = delete;
-
-  ~ThreadPool() {
-    {
-      absl::MutexLock l(mu_);
-      for (size_t i = 0; i < threads_.size(); i++) {
-        queue_.push(nullptr);  // Shutdown signal.
-      }
-    }
-    for (auto &thread : threads_) thread.join();
-  }
-
-  void Schedule(absl::AnyInvocable<void()> func) {
-    absl::MutexLock l(mu_);
-    queue_.push(std::move(func));
-  }
-
-  size_t num_threads() const { return threads_.size(); }
-
- private:
-  bool WorkAvailable() const ABSL_EXCLUSIVE_LOCKS_REQUIRED(mu_) {
-    return !queue_.empty();
-  }
-
-  void WorkLoop() {
-    while (true) {
-      absl::AnyInvocable<void()> func;
-      {
-        absl::MutexLock l(mu_);
-        mu_.Await(absl::Condition(this, &ThreadPool::WorkAvailable));
-        func = std::move(queue_.front());
-        queue_.pop();
-      }
-      if (func == nullptr) {  // Shutdown signal.
-        break;
-      }
-      func();
-    }
-  }
-
-  absl::Mutex mu_;
-  std::queue<absl::AnyInvocable<void()>> queue_ ABSL_GUARDED_BY(mu_);
-  std::vector<std::thread> threads_;
-};
-
 namespace log_domain {
 
 double LogSum(const std::vector<double> &xs);
